+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# v1.08
+* Added a rescue for Errno::ECONNRESET which is usually caused by a badly-behaved client. "Connection reset by peer" is the TCP/IP equivalent of slamming the phone down on the hook. It's more polite than merely not replying, leaving one hanging, but it's not the FIN-ACK expected of the truly polite TCP/IP conversation.
+* Removed `set_mail_id` and replaced it with the Process::pid in parenthesis ater the log level. This is because one needs to know which message is part of a set when more than one process is adding messages to the log. In this example,
+```
+2015-11-13 03:58:11 [INFO] (022999) Connection accepted on port 25 from port 33402 at ::ffff:166.78.151.141 (rubymta-test)
+2015-11-13 03:58:11 [WARN] (023004) Connection failure on port 587 ignored; probably caused by a port scan
+2015-11-13 03:58:11 [INFO] (022999) MySQL database czar_development opened on localhost by czar
+2015-11-13 03:58:11 [INFO] (022999) Receiving message id: 1Zx5V5-1Kefa1-C4
+```
+there is a message for PID 023004 mixed in with the messages for PID 022999.
+
 # v1.07
 * Rescues added for two errors caused by client closing port early.
 * Corrected some text in a comment.